Add HTTPS support for ellis
By default, nothing is changed, and ellis will listen on port 80.

If an SSL key and certificate are placed in /etc/nginx/ssl/ellis.{key,crt},
then nginx will be configured to provide HTTPS to ellis on port 443 and to
redirect from HTTP on port 80 to HTTPS.

The SSL certificate should have a CN of ellis.$home_domain, or
ellis_hostname should be set in the clearwater config to a valid CN
for the certificate.